Add maven support and remove AI support from app
Tess4J (OCR AI) is a legacy lib that doesn't support modules, and it's been impossible to make it work after trying Maven and Gradle (which is unstable now with Java 21), error after errors, and warnings. The changes in this branch are not official, and only useful for generating runtime images or packaging the app.